Update fidget directory name
