+++ src/fidgets/index.ts
@@ -1,5 +1,5 @@
 // TO DO: Figure out how to do this importing dynamically
-import ChannelFidget from "./example";
+import ChannelFidget from "./channel";
 
 export default {
   example: ChannelFidget,
